pyproject.toml
[project]
name = "rgbtrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations==1.4.2",
    "bokeh==3.4.0",
    "clip",
    "colorama==0.4.6",
    "einops==0.7.0",
    "fvcore==0.1.5.post20221221",
    "g4f==0.2.7.1",
    "gputil==1.4.0",
    "h5py==3.10.0",
    "hydra-core>=1.3.2",
    "imageio==2.34.0",
    "imgaug==0.4.0",
    "ipywidgets==8.1.2",
    "joblib==1.3.2",
    "jupyterlab==4.1.5",
    "kornia==0.7.2",
    "meshcat==0.3.2",
    "ninja==1.11.1.1",
    "nodejs==0.1.1",
    "numba==0.59.1",
    "numpy==1.26.4",
    "objaverse==0.1.7",
    "omegaconf==2.3.0",
    "open3d==0.18.0",
    "opencv-contrib-python==4.9.0.80",
    "opencv-python==4.9.0.80",
    "openpyxl==3.1.2",
    "panda3d==1.10.14",
    "pin==2.7.0",
    "plotly==5.20.0",
    "py-spy==0.3.14",
    "pybind11==2.12.0",
    "pybullet==3.2.6",
    "pycocotools==2.0.7",
    "pydantic>=2.12.5",
    "pyglet==1.5.28",
    "pyopengl>=3.1.0",
    "pyopengl-accelerate>=3.1.0",
    "pypng==0.20220715.0",
    "pyrender==0.1.45",
    "pysdf==0.1.9",
    "pyyaml==6.0.1",
    "roma==1.4.4",
    "rtree==1.2.0",
    "ruamel-yaml==0.18.6",
    "scikit-image==0.22.0",
    "scikit-learn==1.4.1.post1",
    "scipy==1.12.0",
    "seaborn==0.13.2",
    "simplejson==3.19.2",
    "timm==0.9.16",
    "torch>=2.0.0",
    "torchvision>=0.15.1",
    "transformations==2024.6.1",
    "trimesh==4.2.2",
    "ultralytics==8.0.120",
    "videoio==0.2.8",
    "wandb==0.16.5",
    "warp-lang==1.0.2",
    "webdataset==0.2.86",
    "xatlas==0.0.9",
    "xlsxwriter==3.2.0",
    "gradio>=4.0.0",
    "pyzmq>=25.0.0",
    "autopep8>=2.3.2",
    "mvsdk",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
clip = { git = "https://github.com/openai/CLIP.git" }
mvsdk = { workspace = true }

[tool.uv.workspace]
members = [
    "third_party/mvsdk-python",
]

[tool.autopep8]
ignore = ["E401", "E402"]
max_line_length = 200
